feat(GCP): progress on GCP call via Validator
