+++ composer.json
@@ -61,13 +61,6 @@
         "psr-4": {
             "App\\": "src/",
             "App\\Application\\": "src/Application/",
-            "App\\Application\\Bridge\\": "src/Application/Bridge/",
-            "App\\Application\\Bridge\\Interfaces\\": "src/Application/Bridge/Interfaces/",
-            "App\\Application\\Helper\\": "src/Application/Helper/",
-            "App\\Application\\Helper\\CloudStorage\\": "src/Application/Helper/CloudStorage/",
-            "App\\Application\\Helper\\CloudStorage\\Interfaces\\": "src/Application/Helper/CloudStorage/Interfaces/",
-            "App\\Application\\Helper\\CloudVision\\": "src/Application/Helper/CloudVision/",
-            "App\\Application\\Helper\\CloudVision\\Interfaces\\": "src/Application/Helper/CloudVision/Interfaces/",
             "App\\Application\\Helper\\Image\\": "src/Application/Helper/Image/",
             "App\\Application\\Helper\\Image\\Interfaces\\": "src/Application/Helper/Image/Interfaces/",
             "App\\Application\\Helper\\Security\\": "src/Application/Helper/Security/",
@@ -96,6 +89,12 @@
             "App\\Infra\\Form\\FormHandler\\": "src/Infra/Form/FormHandler/",
             "App\\Infra\\Form\\FormSubscriber\\": "src/Infra/Form/FormSubscriber/",
             "App\\Infra\\Form\\FormSubscriber\\Interfaces\\": "src/Infra/Form/FormSubscriber/Interfaces/",
+            "App\\Infra\\GCP\\Bridge\\": "src/Infra/GCP/Bridge/",
+            "App\\Infra\\GCP\\Bridge\\Interfaces\\": "src/Infra/GCP/Bridge/Interfaces/",
+            "App\\Infra\\GCP\\CloudStorage\\": "src/Infra/GCP/CloudStorage/",
+            "App\\Infra\\GCP\\CloudStorage\\Interfaces\\": "src/Infra/GCP/CloudStorage/Interfaces/",
+            "App\\Infra\\GCP\\CloudVision\\": "src/Infra/GCP/CloudVision/",
+            "App\\Infra\\GCP\\CloudVision\\Interfaces\\": "src/Infra/GCP/CloudVision/Interfaces/",
             "App\\UI\\Action\\Core\\": "src/UI/Action/Core/",
             "App\\UI\\Action\\Dashboard\\": "src/UI/Action/Dashboard/",
             "App\\UI\\Action\\Security\\": "src/UI/Action/Security/",
@@ -124,7 +123,9 @@
             "App\\Tests\\Form\\Type\\": "tests/Form/Type/",
             "App\\Tests\\Infra\\Form\\FormHandler\\": "tests/Infra/Form/FormHandler/",
             "App\\Tests\\Infra\\Form\\FormSubscriber\\": "tests/Infra/Form/FormSubscriber/",
-            "App\\Tests\\Infra\\Helper\\Security\\": "tests/Infra/Helper/Security/",
+            "App\\Tests\\Infra\\GCP\\Bridge\\": "tests/Infra/GCP/Bridge/",
+            "App\\Tests\\Infra\\GCP\\CloudStorage\\": "tests/Infra/GCP/CloudStorage/",
+            "App\\Tests\\Infra\\GCP\\CloudVision\\": "tests/Infra/GCP/CloudVision/",
             "App\\Tests\\UI\\Action\\Core\\": "tests/UI/Action/Core/",
             "App\\Tests\\UI\\Action\\Dashboard\\": "tests/UI/Action/Dashboard/",
             "App\\Tests\\UI\\Action\\Security\\": "tests/UI/Action/Security/",
