--- src/Application/Bridge/Interfaces/CloudVisionBridgeInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the MarketReminder project.
- *
- * (c) Guillaume Loulier <contact@guillaumeloulier.fr>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace App\Application\Bridge\Interfaces;
-
-/**
- * Interface CloudVisionBridgeInterface.
- *
- * @author Guillaume Loulier <contact@guillaumeloulier.fr>
- */
-interface CloudVisionBridgeInterface extends CloudBridgeInterface
-{
-}
