--- src/Application/Helper/CloudVision/CloudVisionAnalyserHelper.php
@@ -1,56 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the MarketReminder project.
- *
- * (c) Guillaume Loulier <contact@guillaumeloulier.fr>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace App\Application\Helper\CloudVision;
-
-use App\Application\Bridge\Interfaces\CloudVisionBridgeInterface;
-use App\Application\Helper\CloudVision\Interfaces\CloudVisionAnalyserHelperInterface;
-use Google\Cloud\Vision\Image;
-
-/**
- * Class CloudVisionAnalyserHelper.
- *
- * @author Guillaume Loulier <contact@guillaumeloulier.fr>
- */
-class CloudVisionAnalyserHelper implements CloudVisionAnalyserHelperInterface
-{
-    /**
-     * @var CloudVisionBridgeInterface
-     */
-    private $cloudVisionBridgeInterface;
-
-    /**
-     * CloudVisionAnalyserHelper constructor.
-     *
-     * @param CloudVisionBridgeInterface $cloudVisionBridgeInterface
-     */
-    public function __construct(CloudVisionBridgeInterface $cloudVisionBridgeInterface)
-    {
-        $this->cloudVisionBridgeInterface = $cloudVisionBridgeInterface;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function analyse(string $imagePath, string $analyseMode): Image
-    {
-        return $this->cloudVisionBridgeInterface
-                    ->loadCredentialsFile()
-                    ->getServiceBuilder()
-                    ->vision()
-                    ->image(
-                        file_get_contents($imagePath),
-                        [$analyseMode]
-                    );
-    }
-}
