+++ src/Infra/GCP/Bridge/AbstractBridge.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the MarketReminder project.
+ *
+ * (c) Guillaume Loulier <contact@guillaumeloulier.fr>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Infra\GCP\Bridge;
+
+use App\Infra\GCP\Bridge\Interfaces\CloudBridgeInterface;
+use Google\Cloud\Core\ServiceBuilder;
+
+/**
+ * Class AbstractBridge.
+ *
+ * @author Guillaume Loulier <contact@guillaumeloulier.fr>
+ */
+abstract class AbstractBridge implements CloudBridgeInterface
+{
+    /**
+     * @var array
+     */
+    protected $credentials;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getServiceBuilder(): ServiceBuilder
+    {
+        return new ServiceBuilder([
+            'keyFile' => $this->credentials
+        ]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCredentials():? array
+    {
+        return $this->credentials;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function closeConnexion(): void
+    {
+        $this->credentials = null;
+    }
+}
