+++ src/Infra/GCP/Bridge/Interfaces/CloudStorageBridgeInterface.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the MarketReminder project.
+ *
+ * (c) Guillaume Loulier <contact@guillaumeloulier.fr>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Infra\GCP\Bridge\Interfaces;
+
+/**
+ * Interface CloudStorageBridgeInterface.
+ *
+ * @author Guillaume Loulier <contact@guillaumeloulier.fr>
+ */
+interface CloudStorageBridgeInterface extends CloudBridgeInterface
+{
+}
